src/dvc_pipeline/requirements.txt
imbalanced_learn==0.14.0
imblearn==0.0
keras==3.11.3
lightgbm==4.6.0
matplotlib==3.10.6
mlflow==3.2.0
mlflow_skinny==3.2.0
mlflow_tracing==3.2.0
nltk==3.9.1
optuna==4.4.0
pandas==2.3.3
plotly==6.2.0
PyYAML==6.0.2
regex==2024.11.6
scikit_learn==1.7.2
seaborn==0.13.2
tensorflow==2.20.0
xgboost==3.0.5
boto3
dvc[s3]
kaleido